refactor: Update 'estado' field default value in 'Postulacion' model
The 'estado' field in the 'Postulacion' model of 'api/models.py' has been updated to have a default value of `False` instead of `True`. This change ensures that the default state of a postulation is now set to "not active".

Note: Please remove any meta information like issue references, tags, or author names from the commit message.